Exp Calculation Update
The exp is calculateed using temporary exp and not total exp

+++ DesktopIdleProject/Assets/Scripts/Data/Player/PlayerFisher.cs
@@ -87,7 +87,7 @@ public class PlayerFisher : Player
         // animation
         animator.SetTrigger("Caught");
 
-        int rewardedExp;
+        long rewardedExp;
 
         // Fish caught
         bool hasAlreadyFish = PlayerManager.Instance.Inventory.HasItem(hookedFish.Id);
