+++ DesktopIdleProject/Assets/Scripts/SaveSystem/Player/PlayerMinerSaveData.cs
@@ -15,7 +15,7 @@ public class PlayerMinerSaveData
     // ---- STAT VALUES
 
     public int currentLevel;
-    public int currentExp;
+    public long currentExp;
 
     // ---- WEAPON
